fix: solidity <0.8.4 incompatible with custom errors + formatting

+++ src/bridges/liquity/interfaces/IBorrowerOperations.sol
@@ -2,11 +2,23 @@
 pragma solidity >=0.8.0 <=0.8.10;
 
 interface IBorrowerOperations {
-    function openTrove(uint _maxFee, uint _LUSDAmount, address _upperHint, address _lowerHint) external payable;
+    function openTrove(
+        uint256 _maxFee,
+        uint256 _LUSDAmount,
+        address _upperHint,
+        address _lowerHint
+    ) external payable;
 
     function closeTrove() external;
 
-    function adjustTrove(uint _maxFee, uint _collWithdrawal, uint _debtChange, bool isDebtIncrease, address _upperHint, address _lowerHint) external payable;
+    function adjustTrove(
+        uint256 _maxFee,
+        uint256 _collWithdrawal,
+        uint256 _debtChange,
+        bool isDebtIncrease,
+        address _upperHint,
+        address _lowerHint
+    ) external payable;
 
     function claimCollateral() external;
 }
