+++ src/bridges/liquity/interfaces/IPriceFeed.sol
@@ -2,5 +2,5 @@
 pragma solidity >=0.8.0 <=0.8.10;
 
 interface IPriceFeed {
-    function fetchPrice() external returns (uint);
+    function fetchPrice() external returns (uint256);
 }
