+++ src/bridges/liquity/interfaces/ITroveManager.sol
@@ -1,39 +1,41 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.8.0 <=0.8.10;
 
-import "./ILiquityBase.sol";
-import "./IStabilityPool.sol";
-import "./ILQTYStaking.sol";
-
+import './ILiquityBase.sol';
+import './IStabilityPool.sol';
+import './ILQTYStaking.sol';
 
 // Common interface for the Trove Manager.
 interface ITroveManager is ILiquityBase {
-    function getCurrentICR(address _borrower, uint _price) external view returns (uint);
+    function getCurrentICR(address _borrower, uint256 _price) external view returns (uint256);
 
     function liquidate(address _borrower) external;
 
     function redeemCollateral(
-        uint _LUSDAmount,
+        uint256 _LUSDAmount,
         address _firstRedemptionHint,
         address _upperPartialRedemptionHint,
         address _lowerPartialRedemptionHint,
-        uint _partialRedemptionHintNICR,
-        uint _maxIterations,
-        uint _maxFee
+        uint256 _partialRedemptionHintNICR,
+        uint256 _maxIterations,
+        uint256 _maxFee
     ) external;
 
-    function getEntireDebtAndColl(address _borrower) external view returns (
-        uint debt, 
-        uint coll, 
-        uint pendingLUSDDebtReward, 
-        uint pendingETHReward
-    );
+    function getEntireDebtAndColl(address _borrower)
+        external
+        view
+        returns (
+            uint256 debt,
+            uint256 coll,
+            uint256 pendingLUSDDebtReward,
+            uint256 pendingETHReward
+        );
 
     function closeTrove(address _borrower) external;
 
-    function getBorrowingRateWithDecay() external view returns (uint);
+    function getBorrowingRateWithDecay() external view returns (uint256);
 
-    function getTroveStatus(address _borrower) external view returns (uint);
-    
-    function checkRecoveryMode(uint _price) external view returns (bool);
+    function getTroveStatus(address _borrower) external view returns (uint256);
+
+    function checkRecoveryMode(uint256 _price) external view returns (bool);
 }
